remove background and refactor the fetch data
