+++ manifest.json
@@ -11,6 +11,5 @@
     "default_popup": "banano.html",
     "default_icon": "icon.png"
   },
-  "background": [{ "scripts": "background.js" }],
   "manifest_version": 2
 }
